レイアウトを使わずコンポーネントを使ってイベントデータを取得するよう修正 #11
