--- resources/ts/util/context.ts
@@ -1,14 +1,0 @@
-import React from "react";
-import { Colors, GraphData, TotalData } from "@/types";
-
-type GraphContextType = {
-    graphData: GraphData;
-    colors: Colors;
-    total: TotalData;
-};
-
-export const GraphDataContext = React.createContext<GraphContextType>({
-    graphData: {},
-    colors: [],
-    total: {},
-});
